refactor Image, ImageIO     - ImageIO.bmp, NeHeOGL_ImageIO_bmp coding refactor     - NeHeOGLImageIO error handle with error_longjmp_restore_point     - and some minor update

+++ src/tutorials/NeHe OpenGL/src/NeHeOGL_Image.cpp
@@ -58,8 +58,8 @@ void NeHeOGL_Image::create(ColorType colorType, int width, int height, int strid
 	_create(colorType, width, height, strideInBytes, 1, strideInBytes * height);
 }
 
-void NeHeOGL_Image::create(ColorType colorType, int width, int height, int strideInBytes, int mipmapCount, int dataSizeInBytes) {
-	_create(colorType, width, height, strideInBytes, mipmapCount, static_cast<size_t>(dataSizeInBytes));
+void NeHeOGL_Image::create(ColorType colorType, int width, int height, int strideInBytes, int mipmapCount, size_t dataSizeInBytes) {
+	_create(colorType, width, height, strideInBytes, mipmapCount, dataSizeInBytes);
 }
 
 void NeHeOGL_Image::_create(ColorType colorType, int width, int height, int strideInBytes, int mipmapCount, size_t dataSizeInBytes) {
